+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Note: Spacedock's markdown doesn't recognize lists using `-`, so make sure to
 -->
 
 ## Unreleased
+## Fixed
+* Fixed a bug where empty SRBs would occasionally keep running in the background.
+
 ## 0.6.6
 ### Fixed
 * Fixed an incompatibility due to different types between MechJeb 2.14.3 and 2.15.