+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Note: Spacedock's markdown doesn't recognize lists using `-`, so make sure to
 ## Fixed
 * Fixed a bug where empty SRBs would occasionally keep running in the background.
 * Fixed logspam due to missing simulation resources.
+* Simplify the code that slows down time warps when executing maneuver nodes.
 
 ## 0.6.6
 ### Fixed